build(twig-refactoring) Implementation of Twig templates and global refactoring
