+++ app/Controllers/Errors/404.php
@@ -1,0 +1,2 @@
+<?php
+require(ROOT . '../app/Views/Errors/404.php');
