--- app/Controllers/error404.php
@@ -1,2 +1,0 @@
-<?php
-require(dirname(__DIR__) . '../Views/error404.php');
