--- app/Models/GlobalQuery/PostGlobalQuery.php
@@ -1,12 +1,0 @@
-<?php
-namespace App\Models\GlobalQuery;
-
-use App\Models\GlobalQuery\GlobalQuery;
-use App\Models\Post;
-
-final class PostGlobalQuery extends GlobalQuery {
-    
-    protected $table = "article";
-    protected $class = Post::class;
-
-}
