+++ public/index.php
@@ -1,11 +1,13 @@
 <?php
+define('ROOT', dirname(__DIR__));
+
 require '../vendor/autoload.php';
 
 $whoops = new \Whoops\Run;
 $whoops->pushHandler(new \Whoops\Handler\PrettyPageHandler);
 $whoops->register();
 
-$router = new App\Core\Router(dirname(__DIR__) . '/app/Controllers');
+$router = new App\Core\Router(ROOT . '../app/Controllers');
 $router
     // Home
     ->get('/', '/home', 'home')
@@ -13,4 +15,5 @@ $router
     ->get('/articles', '/posts/index', 'articles')
     ->get('/article/[i:id]', '/posts/post', 'article')
     ->post('/article/update/[i:id]', 'posts/update', 'update_article')
+    // Run
     ->run();
