+++ templates/posts/index.html.twig
@@ -1,0 +1,5 @@
+{% extends "base.html.twig" %}
+
+{% block body %}
+    <h1>Hello World</h1>
+{% endblock %}
